Move investor icons from index-page folder to shared data folder, slice the rendered founding members array
