+++ src/data/investors/basicsCapital.js
