+++ src/data/investors/bitmain.js
