+++ src/data/investors/dcg.js
