+++ src/data/investors/gateIOLabs.js
