+++ src/data/investors/genblockCapital.js
