+++ src/data/investors/gsr.js
