+++ src/data/investors/hypersphere.js
