+++ src/data/investors/kernelVentures.js
