+++ src/data/investors/ldCapital.js
