+++ src/data/investors/m6.js
