+++ src/data/investors/newTribeCapital.js
