+++ src/data/investors/pandaCapital.js
