+++ src/data/investors/sneakyVC.js
